requirements.txt
certifi==2018.4.16
future==0.16.0
google-api-python-client==1.6.5
httplib2==0.11.0
oauth2client==4.1.2
pkg-resources==0.0.0
psycopg2==2.7.4
pyasn1==0.4.2
pyasn1-modules==0.2.1
python-telegram-bot==10.0.2
rsa==3.4.2
simpledbf==0.2.6
six==1.11.0
uritemplate==3.0.0
